add: snackbar_manager -> snack_bar_helperに変更・utils/helperにファイルを移動しました
