--- lib/core/utils/snack_bar_manager.dart
@@ -1,66 +1,0 @@
-import 'package:awesome_snackbar_content/awesome_snackbar_content.dart';
-import 'package:flutter/material.dart';
-
-void openErrorSnackBar(
-  BuildContext context,
-  String title,
-  String message,
-) {
-  final snackBar = SnackBar(
-    elevation: 0,
-    behavior: SnackBarBehavior.floating,
-    backgroundColor: Colors.transparent,
-    duration: const Duration(seconds: 5),
-    content: AwesomeSnackbarContent(
-      title: title,
-      message: message,
-      contentType: ContentType.failure,
-      titleTextStyle: TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-      messageTextStyle: TextStyle(fontSize: 12),
-    ),
-  );
-  ScaffoldMessenger.of(context)
-    ..hideCurrentSnackBar()
-    ..showSnackBar(snackBar);
-}
-
-void openSuccessSnackBar(
-  BuildContext context,
-  String title,
-  String message,
-) {
-  final snackBar = SnackBar(
-    elevation: 0,
-    behavior: SnackBarBehavior.floating,
-    backgroundColor: Colors.transparent,
-    content: AwesomeSnackbarContent(
-      title: title,
-      message: message,
-      contentType: ContentType.success,
-    ),
-  );
-  ScaffoldMessenger.of(context)
-    ..hideCurrentSnackBar()
-    ..showSnackBar(snackBar);
-}
-
-void openComingSoonSnackBar(BuildContext context) {
-  ScaffoldMessenger.of(context).showSnackBar(
-    SnackBar(
-      content: Center(
-        child: Text(
-          '🙇　Coming Soon　🙇',
-          style: TextStyle(
-            fontWeight: FontWeight.bold,
-            color: Colors.white,
-            fontSize: 20,
-          ),
-        ),
-      ),
-    ),
-  );
-}
-
-void hideSnackBar(BuildContext context) {
-  ScaffoldMessenger.of(context).hideCurrentSnackBar();
-}
