+++ lib/ui/screen/authentication/authentication_screen.dart
@@ -6,8 +6,8 @@ import 'package:auth_buttons/auth_buttons.dart'
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:food_gram_app/core/data/supabase/auth/account_service.dart';
+import 'package:food_gram_app/core/utils/helpers/snack_bar_helper.dart';
 import 'package:food_gram_app/core/utils/provider/loading.dart';
-import 'package:food_gram_app/core/utils/snack_bar_manager.dart';
 import 'package:food_gram_app/gen/assets.gen.dart';
 import 'package:food_gram_app/gen/l10n/l10n.dart';
 import 'package:food_gram_app/main.dart';
@@ -139,7 +139,7 @@ class AuthenticationScreen extends HookConsumerWidget {
                         if (Platform.isIOS) {
                           controller.loginApple(context);
                         } else {
-                          openErrorSnackBar(
+                          SnackBarHelper().openErrorSnackBar(
                             context,
                             L10n.of(context).appleLoginFailure,
                             '',
@@ -174,7 +174,7 @@ class AuthenticationScreen extends HookConsumerWidget {
   }
 
   Future<void> redirect(BuildContext context, WidgetRef ref) async {
-    hideSnackBar(context);
+    SnackBarHelper().hideSnackBar(context);
     if (!await AccountService.isUserRegistered()) {
       context.pushReplacementNamed(RouterPath.tab);
     } else {
