+++ lib/ui/screen/my_profile/my_profile_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:food_gram_app/core/data/purchase/subscription_provider.dart';
 import 'package:food_gram_app/core/data/supabase/post_stream.dart';
 import 'package:food_gram_app/core/utils/async_value_group.dart';
-import 'package:food_gram_app/core/utils/snack_bar_manager.dart';
+import 'package:food_gram_app/core/utils/helpers/snack_bar_helper.dart';
 import 'package:food_gram_app/router/router.dart';
 import 'package:food_gram_app/ui/component/app_app_bar.dart';
 import 'package:food_gram_app/ui/component/app_floating_button.dart';
@@ -70,7 +70,7 @@ class MyProfileScreen extends ConsumerWidget {
                       'exchange point',
                       const Duration(seconds: 1),
                       () async {
-                        openComingSoonSnackBar(context);
+                        SnackBarHelper().openComingSoonSnackBar(context);
                       },
                     );
                   },
